# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Dwar Rater API
  description: |
    API for character analysis, snapshots, clan management, and closed profile monitoring
    for the browser MMORPG "Legend: Legacy of the Dragons" (dwar.ru).

    **Authentication:** HTTP Basic Auth. All endpoints except `/api/login`, `/api/health`, and `/` require authentication.

    **Versioning:** canonical version source is the repo-root `VERSION` file (sync `frontend/package.json` on release; keep this schema version in sync manually — the OpenAPI schema is static).
  version: "1.0.0"
  contact:
    name: Dwar Rater
    url: https://github.com/kirillbarashkov/dwar-rater
servers:
  - url: http://localhost:5000
    description: Local development
security:
  - BasicAuth: []
paths:
...
  /api/tracks/generate:
    post:
      summary: Generate track
      description: Generate an improvement track from character data and a scenario.
      tags:
        - Tracks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - character_data
                - scenario_id
              properties:
                character_data:
                  type: object
                scenario_id:
                  type: integer
      responses:
        "201":
          description: Track generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  steps:
                    type: array
                  total_progress:
                    type: number
        "400":
          description: Missing required fields
        "404":
          description: Scenario not found
  /api/tracks/{track_id}:
    get:
      summary: Get track
      tags:
        - Tracks
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete track
      tags:
        - Tracks
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Track deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /api/tracks/{track_id}/re-evaluate:
    post:
      summary: Re-evaluate track
      description: Authoritative re-sync of steps against a fresh source character snapshot (target stays frozen).
      tags:
        - Tracks
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                character_data:
                  type: object
                  description: Fresh processed source character (same character the track was created for)
                force_refresh:
                  type: boolean
                  description: Parse source from the user's profile character_url instead (requires analyze.write)
      responses:
        "200":
          description: Updated steps, phases and power_gap
        "400":
          description: character_data missing or not an object
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Legacy track without target snapshot, or character mismatch
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tracks/{track_id}/step/{step_id}:
    put:
      summary: Update track step
      description: Mark a step as completed or update its status.
      tags:
        - Tracks
      parameters:
        - name: track_id
          in: path
          required: true
          schema:
            type: integer
        - name: step_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
      responses:
        "200":
          description: Step updated with new progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  steps:
                    type: array
                  total_progress:
                    type: number
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    StatusResponse:
      type: object
      properties:
        status:
          type: string
...
    TrackSummary:
      type: object
      properties:
        id:
          type: integer
        character_nick:
          type: string
        scenario_id:
          type: integer
        total_progress:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Track:
      allOf:
        - $ref: "#/components/schemas/TrackSummary"
        - type: object
          properties:
            steps:
              type: array
              items:
                type: object
...